The file implementing parsing and evaluation of macros
This is a very preliminary implementation of a macro system.
Unquote-splicing seems not to work, but backquote and unquote both
seem to work.

The macro system builds on the previous system (where macros needed to
be defined in Clojure).  In this system, macros that are defined in
Clojurescript with the same name as Clojure macros take precedence.
Macros are not yet namespaced, and recursive macros have not yet been
tested.